[FIX] mrp: orderpoint code cleanup
stock.warehouse.orderpoint exists in 2 different files in mrp.
This commit moves the code from the stock_warehouse.py file to the stock_orderpoint.py file.
