+++ addons/mrp/models/stock_orderpoint.py
@@ -2,9 +2,10 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import _, api, fields, models
+from odoo.exceptions import ValidationError
+from odoo.tools import split_every
 from odoo.tools.float_utils import float_is_zero
 from odoo.osv.expression import AND
-from dateutil.relativedelta import relativedelta
 
 
 class StockWarehouseOrderpoint(models.Model):
@@ -146,3 +147,22 @@ class StockWarehouseOrderpoint(models.Model):
             ('state', '=', 'draft'),
         ]).action_confirm()
         return super()._post_process_scheduler()
+
+    @api.constrains('product_id')
+    def check_product_is_not_kit(self):
+        domain = [
+            '|', ('product_id', 'in', self.product_id.ids),
+            '&', ('product_id', '=', False),
+            ('product_tmpl_id', 'in', self.product_id.product_tmpl_id.ids),
+            ('type', '=', 'phantom'),
+        ]
+        if self.env['mrp.bom'].search_count(domain, limit=1):
+            raise ValidationError(_("A product with a kit-type bill of materials can not have a reordering rule."))
+
+    def _get_orderpoint_products(self):
+        non_kit_product_ids = []
+        for products in split_every(2000, super()._get_orderpoint_products().ids, self.env['product.product'].browse):
+            kit_product_ids = {k.id for k in self.env['mrp.bom']._bom_find(products, bom_type='phantom')}
+            non_kit_product_ids.extend(id_ for id_ in products.ids if id_ not in kit_product_ids)
+            products.invalidate_recordset()
+        return self.env['product.product'].browse(non_kit_product_ids)
